+++ DeltaDNA/DDNAEngagement.m
@@ -59,5 +59,20 @@
     };
 }
 
+- (void)setRaw:(NSString *)raw
+{
+    if (raw != nil && raw.length > 0) {
+        NSData * data = [raw dataUsingEncoding:NSUTF8StringEncoding];
+        NSError * error = nil;
+        NSDictionary * result = [NSJSONSerialization JSONObjectWithData:data
+                                                                options:kNilOptions
+                                                                  error:&error];
+        self.json = (error == 0) ? result : nil;
+    } else {
+        self.json = nil;
+    }
+    
+    _raw = raw;
+}
 
 @end