+++ Framework/DeltaDNA.h
@@ -36,5 +36,6 @@ FOUNDATION_EXPORT const unsigned char DeltaDNAVersionString[];
 #import <DeltaDNA/DDNATransaction.h>
 #import <DeltaDNA/DDNAEngagement.h>
 #import <DeltaDNA/DDNAPopup.h>
+#import <DeltaDNA/DDNAImageMessage.h>
 
 #endif /* _DELTADNA_ */